Separate method into common.geoserver and use them in layer.geoserver.sld
